Add separate api endpoint for detailed visualization indexing
